Improve behavior
